# Fix text2d view-visibility