--- a/crates/bevy_text/src/text2d.rs
+++ b/crates/bevy_text/src/text2d.rs
@@ -85,6 +85,7 @@ pub fn extract_text2d_sprite(
     windows: Extract<Query<&Window, With<PrimaryWindow>>>,
     text2d_query: Extract<
         Query<(
+            Entity,
             &ViewVisibility,
             &Text,
             &TextLayoutInfo,
@@ -100,7 +101,9 @@ pub fn extract_text2d_sprite(
         .unwrap_or(1.0);
     let scaling = GlobalTransform::from_scale(Vec2::splat(scale_factor.recip()).extend(1.));
 
-    for (view_visibility, text, text_layout_info, anchor, global_transform) in text2d_query.iter() {
+    for (original_entity, view_visibility, text, text_layout_info, anchor, global_transform) in
+        text2d_query.iter()
+    {
         if !view_visibility.get() {
             continue;
         }
@@ -125,8 +128,9 @@ pub fn extract_text2d_sprite(
             }
             let atlas = texture_atlases.get(&atlas_info.texture_atlas).unwrap();
 
+            let entity = commands.spawn_empty().id();
             extracted_sprites.sprites.insert(
-                commands.spawn_empty().id(),
+                entity,
                 ExtractedSprite {
                     transform: transform * GlobalTransform::from_translation(position.extend(0.)),
                     color,
@@ -138,6 +142,9 @@ pub fn extract_text2d_sprite(
                     anchor: Anchor::Center.as_vec(),
                 },
             );
+            extracted_sprites
+                .original_entities
+                .insert(entity, original_entity);
         }
     }
 }