@rparrett 2023-10-13
> Don’t want to muddy the waters but since the new sprites are only created for visible entities could we just have a force_visible bool that gets set to true on the new ones? Not sure if there’s other value to tracking the original entity

My understanding is that entities that are definitely not visible in any view are not extracted, but the visibility check in `queue_sprites` is per-view / based on `RenderLayers`.